feat: Create a simple API to be deployed using Functions
Create a simple API to be deployed to Azure Functions (compatible with the "Functions" ADE environment).
The QnA API will be used to demonstrate the integration of AZD and ADE via a configurable logic which makes use of "iterations": by defualt we will start with RELEASE=1, and by increasing the release number (to simulate a change) we will get a different behavior which can be increased errors, more latency, or no issues at all.